Починил css, дотипизировал компоненты, поправил мелочи
