+++ .eslintrc.json
@@ -26,9 +26,7 @@
     "linebreak-style": "off",
     "@typescript-eslint/no-var-requires": 0,
     "semi": "off",
-    "@typescript-eslint/semi": "error",
-    "no-use-before-define": 0,
-    "@typescript-eslint/no-use-before-define": 0
+    "@typescript-eslint/semi": "error"
   },
   "settings": {
     "react": {
