+++ src/common/sectionHeader/SectionHeader.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import PropTypes from 'prop-types';
 
 const Headline = styled.h2`
-  font-family: Montserrat;
+  
   font-style: normal;
   font-weight: 600;
   font-size: 30px;
