+++ src/components/Group/Group.jsx
@@ -31,7 +31,7 @@ export default function News() {
 }
 
 const Wrapper = styled.div`
-  font-family: Montserrat;
+  
   background: #111111;
   padding: 100px;
 `;
@@ -55,7 +55,7 @@ const Container = styled.div`
 const Label = styled.div`
   position: absolute;
   top: -90px;
-  font-family: Montserrat;
+  
   font-style: normal;
   font-weight: 600;
   font-size: 32px;
@@ -84,7 +84,7 @@ const DataContainer = styled.div`
 `;
 
 const NameGroup = styled.div`
-  font-family: Montserrat;
+  
   font-style: normal;
   font-weight: 500;
   font-size: 24px;
@@ -95,7 +95,7 @@ const NameGroup = styled.div`
 `;
 
 const Category = styled.div`
-  font-family: Montserrat;
+  
   font-style: normal;
   font-weight: 500;
   font-size: 18px;
