+++ src/components/Main/Articles/blockComment/Comment/index.js
@@ -1,0 +1,3 @@
+import Comment from './Comment.tsx';
+
+export default Comment;
