+++ src/components/Main/Articles/blockComment/Comment/styles.js
@@ -1,0 +1,39 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div``;
+export const UserInfo = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+export const TitleWrap = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  margin-left: 25px;
+`;
+export const Title = styled.h3`
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 20px;
+  margin: 0 0 5px 0;
+  color: black;
+  font-family: Montserrat, serif;
+`;
+export const SubtitleTime = styled.span`
+  font-weight: 500;
+  font-size: 13px;
+  line-height: 16px;
+  color: #515151;
+`;
+
+export const Text = styled.p`
+  width: 95%;
+  font-size: 16px;
+  line-height: 165%;
+  margin: 0 0 40px 0;
+  color: #000000;
+  text-align: start;
+  padding-left: 95px;
+`;
