+++ src/components/Main/Articles/blockNotes/styles.js
@@ -15,7 +15,6 @@ border-bottom: 1px solid #515151;
 `;
 export const Menu = styled.ul`
 display: flex;
-justify-content: flex-start;
 align-items: center;
 padding: 0;
 margin-bottom: 61px;
@@ -55,11 +54,9 @@ width: 200px;
 height: 35px;
 background: #ffffff;
 border: 1px solid #515151;
-box-sizing: border-box;
 border-radius: 10px;
 font-size: 16px;
 line-height: 20px;
 color: #515151;
 padding: 10px;
-box-sizing: border-box;
 `;
