+++ src/components/News/NewsItem.jsx
@@ -119,7 +119,7 @@ const AuthorContainer = styled.div`
 
 const Author = styled.span`
   display: block;
-  font-family: Montserrat;
+  
   font-style: normal;
   font-weight: 600;
   font-size: 16px;
@@ -131,7 +131,7 @@ const Author = styled.span`
 
 const Time = styled.span`
   display: block;
-  font-family: Montserrat;
+  
   font-style: normal;
   font-weight: 500;
   font-size: 13px;
@@ -151,7 +151,7 @@ const ActionsContainer = styled.div`
 const ButtonAction = styled.button`
   margin-left: 65px;
   display: flex;
-  font-family: Montserrat;
+  
   font-style: normal;
   font-weight: normal;
   font-size: 18px;
