+++ src/index.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-use-before-define */
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
