+++ tsconfig.json
@@ -22,5 +22,9 @@
   },
   "include": [
     "src"
+  ],
+  "exclude": [
+    "*.js",
+    "*.jsx"
   ]
 }
